pyproject.toml
[project]
name = "generative-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.13",
    "langchain-mistralai>=1.1.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
